+++ include/RenderEngine/vec3.hpp
@@ -3,11 +3,11 @@
 
 namespace RenderEngine {
 struct vec3 {
-  double x_;
-  double y_;
-  double z_;
+  float x_;
+  float y_;
+  float z_;
   
- vec3(double x, double y, double z) noexcept;
+ vec3(float x, float y, float z) noexcept;
  vec3() noexcept;
 
   [[nodiscard]] static vec3 zero() noexcept;
@@ -26,26 +26,26 @@ struct vec3 {
  vec3 &operator-=(const vec3 &rhs) noexcept;
  vec3 &operator+=(const vec3 &rhs) noexcept;
 
-  [[nodiscard]] vec3 operator/(double rhs) const noexcept;
+  [[nodiscard]] vec3 operator/(float rhs) const noexcept;
 
- vec3 &operator*=(double rhs) noexcept;
- vec3 &operator/=(double rhs) noexcept;
+ vec3 &operator*=(float rhs) noexcept;
+ vec3 &operator/=(float rhs) noexcept;
 
-  [[nodiscard]] double lengthSquared() const noexcept;
-  [[nodiscard]] double length() const noexcept;
+  [[nodiscard]] float lengthSquared() const noexcept;
+  [[nodiscard]] float length() const noexcept;
 
  vec3 &normalise();
 
- void update(double x, double y, double z) noexcept;
+ void update(float x, float y, float z) noexcept;
 };
 
 [[nodiscard]] vec3 crossProduct(const vec3 &a,
                                           const vec3 &b) noexcept;
-[[nodiscard]] double dotProduct(const vec3 &a,
+[[nodiscard]] float dotProduct(const vec3 &a,
                                           const vec3 &b) noexcept;
 
-[[nodiscard]] vec3 operator*(double scalar, const vec3 &vec) noexcept;
-[[nodiscard]] vec3 operator*(const vec3 &vec, double scalar) noexcept;
+[[nodiscard]] vec3 operator*(float scalar, const vec3 &vec) noexcept;
+[[nodiscard]] vec3 operator*(const vec3 &vec, float scalar) noexcept;
 
 } // namespace RenderEngine
