+++ src/Light.cpp
@@ -2,6 +2,6 @@
 
 namespace RenderEngine {
 
-Light::Light(double intensity, const vec3 &position) noexcept
+Light::Light(float intensity, const vec3 &position) noexcept
     : intensity_(intensity), position_(position) {}
 } // namespace RenderEngine
