+++ src/vec2.cpp
@@ -3,10 +3,10 @@
 #include <stdexcept>
 
 namespace RenderEngine {
- vec2::vec2(double x, double y) noexcept : x_(x), y_(y) {}
- vec2::vec2() noexcept : x_(0.), y_(0.) {}
+ vec2::vec2(float x, float y) noexcept : x_(x), y_(y) {}
+ vec2::vec2() noexcept : x_(0.f), y_(0.f) {}
 
-[[nodiscard]] vec2 vec2::zero() noexcept { return vec2(0., 0.); };
+[[nodiscard]] vec2 vec2::zero() noexcept { return vec2(0.f, 0.f); };
 
 [[nodiscard]] bool vec2::operator==(const vec2 &rhs) const noexcept {
   return x_ == rhs.x_ && y_ == rhs.y_;
@@ -36,40 +36,40 @@ namespace RenderEngine {
   return *this;
 }
 
-[[nodiscard]] vec2 operator*(double scalar, const vec2 &vec) noexcept {
+[[nodiscard]] vec2 operator*(float scalar, const vec2 &vec) noexcept {
   return vec2(vec.x_ * scalar, vec.y_ * scalar);
 }
 
-[[nodiscard]] vec2 operator*(const vec2 &vec, double scalar) noexcept {
+[[nodiscard]] vec2 operator*(const vec2 &vec, float scalar) noexcept {
   return vec2(vec.x_ * scalar, vec.y_ * scalar);
 }
 
-[[nodiscard]] vec2 vec2::operator/(double rhs) const noexcept {
+[[nodiscard]] vec2 vec2::operator/(float rhs) const noexcept {
   return vec2(x_ / rhs, y_ / rhs);
 }
 
- vec2 &vec2::operator*=(double rhs) noexcept {
+ vec2 &vec2::operator*=(float rhs) noexcept {
   x_ *= rhs;
   y_ *= rhs;
   return *this;
 }
 
- vec2 &vec2::operator/=(double rhs) noexcept {
+ vec2 &vec2::operator/=(float rhs) noexcept {
   x_ /= rhs;
   y_ /= rhs;
   return *this;
 }
 
-[[nodiscard]] double vec2::lengthSquared() const noexcept {
+[[nodiscard]] float vec2::lengthSquared() const noexcept {
   return x_ * x_ + y_ * y_;
 }
 
-[[nodiscard]] double vec2::length() const noexcept {
+[[nodiscard]] float vec2::length() const noexcept {
   return std::sqrt(lengthSquared());
 }
 
 [[nodiscard]] vec2 &vec2::normalise() {
-  double length = lengthSquared();
+  float length = lengthSquared();
   if (length == 0) {
     throw std::runtime_error("Cannot normalise zero vector");
   }
@@ -82,12 +82,12 @@ namespace RenderEngine {
   return *this;
 }
 
-[[nodiscard]] double crossProduct(const vec2 &a,
+[[nodiscard]] float crossProduct(const vec2 &a,
                                             const vec2 &b) noexcept {
   return a.x_ * b.y_ - a.y_ * b.x_;
 }
 
-[[nodiscard]] double dotProduct(const vec2 &a,
+[[nodiscard]] float dotProduct(const vec2 &a,
                                           const vec2 &b) noexcept {
   return a.x_ * b.x_ + a.y_ * b.y_;
 }
